Second review: Redis bans expire, server-clock deadlines, config objects, mypy in CI
Checked each point in code first; fixes for the ones that held:
- Redis: bans of keys rejected with 401/403 expire after invalid_ttl
  (default one day) for other workers; every deadline is computed on the
  Redis server clock (a client 1000 s ahead parked a key for 1030 s).
- The first Redis error was not logged during the first 30 s after boot
  (warning throttle started at monotonic 0) - the 'flaky' test.
- Cancellation / Ctrl+C no longer waits for pending Redis updates; they
  are sent from a daemon thread.
- rotator.respect_rate_limit_headers raised AttributeError.
- Config objects: RetryConfig, RateLimitConfig, SharedStateConfig,
  RequestConfig, HTTPConfig (flat arguments keep working).
- mypy: 94 errors -> 0, new CI job gating releases.
- Deprecated RateLimitMiddleware and unused utils helpers (removal in 1.0).

+++ apikeyrotator/core/policy.py
@@ -83,7 +83,7 @@ class RetryPolicy:
     def backoff(self, attempt: int) -> float:
         # Cap the exponent too, to avoid float overflow for huge max_retries
         delay = min(self.base_delay * (2 ** min(attempt, 30)), self.max_delay)
-        return min(delay + random.uniform(0, delay * 0.1), self.max_delay)
+        return float(min(delay + random.uniform(0, delay * 0.1), self.max_delay))
 
     def random_delay(self) -> float:
         low_high = self.random_delay_range
